refactor(unique-id): move persistent ID restoration logic from background into UniqueId module
